+++ FFXIVClientStructs/FFXIV/Client/UI/Agent/LobbyUIClient.cs
@@ -1,10 +1,12 @@
+using FFXIVClientStructs.FFXIV.Application.Network.LobbyClient;
 using FFXIVClientStructs.FFXIV.Client.Network;
 
 namespace FFXIVClientStructs.FFXIV.Client.UI.Agent;
 
 // Client::UI::Agent::LobbyUIClient
 //   Application::Network::LobbyClient::LobbyRequestCallback
-[GenerateInterop]
+[GenerateInterop(isInherited: true)]
+[Inherits<LobbyRequestCallback>]
 [VirtualTable("48 8D 05 ?? ?? ?? ?? ?? ?? ?? 48 8B F9 48 89 71 ?? 48 89 71 ?? 48 89 71 ?? 48 89 71 ?? 48 89 71", 3)]
 [StructLayout(LayoutKind.Explicit, Size = 0x8C8)]
 public unsafe partial struct LobbyUIClient {
@@ -15,10 +17,6 @@ public unsafe partial struct LobbyUIClient {
     [FieldOffset(0x30)] public StdVector<LobbyDataCenterWorldEntry> CurrentDataCenterWorlds;
 
     [FieldOffset(0x48)] public LobbySubscriptionInfo* SubscriptionInfo;
-
-    // Technically part of LobbyRequestCallback
-    [VirtualFunction(4)]
-    public partial void ReportError(LobbyStatusCode* status);
 }
 
 [GenerateInterop]
@@ -40,12 +38,4 @@ public struct LobbySubscriptionInfo // name probably totally wrong
     [FieldOffset(0x30)] public uint TotalDaysSubscribed;
     [FieldOffset(0x34)] public uint DaysRemaining;
     [FieldOffset(0x38)] public uint DaysUntilNextVeteranRank;
-}
-
-[StructLayout(LayoutKind.Explicit, Size = 0x80)]
-public struct LobbyStatusCode {
-    [FieldOffset(0x00)] public int Code;
-    [FieldOffset(0x08)] public int CodeType;
-    [FieldOffset(0x10)] public Utf8String String;
-    [FieldOffset(0x78)] public ushort ErrorSheetRow;
 }
